Refactor based on review feedback
Refactor the lambda handlers and shared functionality into their own
subpackages.

Add some testing to untested shared functions.
