+++ tests/unit/fixtures/account_list.py
@@ -1,0 +1,8 @@
+# mock return for list_accounts()
+mock_account_list = {
+    'Accounts': [
+        { 'Id': '111222333444', },
+        { 'Id': '222333444555', },
+        { 'Id': '333444555666', },
+    ]
+}
