+++ tests/unit/fixtures/tag_names.py
@@ -1,0 +1,5 @@
+# injected env var value
+tag_list_string = 'TagOne,TagTwo'
+
+# expected parsed env var value
+expected_tag_list = [ 'TagOne', 'TagTwo' ]
